make types consistent in chunksize and related loops

+++ contracts/Locker.sol
@@ -25,7 +25,7 @@ contract Locker is Restricted, TokenReceiver {
 
     using SafeMath for uint256;
 
-    uint public constant CHUNK_SIZE = 100;
+    uint16 public constant CHUNK_SIZE = 100;
 
     event NewLockProduct(uint32 indexed lockProductId, uint32 perTermInterest, uint32 durationInSecs,
                             uint32 minimumLockAmount, bool isActive);
@@ -148,7 +148,7 @@ contract Locker is Restricted, TokenReceiver {
     // returns CHUNK_SIZE lock products starting from some offset
     // lock products are encoded as [ perTermInterest, durationInSecs, minimumLockAmount, maxLockAmount, isActive ]
     function getLockProducts(uint offset) external view returns (uint[5][CHUNK_SIZE] response) {
-        for (uint8 i = 0; i < CHUNK_SIZE; i++) {
+        for (uint16 i = 0; i < CHUNK_SIZE; i++) {
 
             if (offset + i >= lockProducts.length) { break; }
